Move SideContentHeight to SideContentReducer
