+++ src/commons/sideContent/SideContentProvider.tsx
@@ -1,3 +1,5 @@
+import { useEffect } from 'react';
+
 import { useSideContent } from './SideContentHelper';
 import type {
   ChangeTabsCallback,
@@ -25,29 +27,50 @@ type SideContentProviderProps = {
    */
   onChange?: ChangeTabsCallback;
   selectedTab?: SideContentType;
+  
+  /**
+   * Value to use if the currently selected tab is undefined
+   */
   defaultTab?: SideContentType;
   workspaceLocation: SideContentLocation;
 };
 
+/**
+ * Component that connects its child SideContentComponent to the Redux store, automatically
+ * providing SideContentHeight information.
+ * 
+ * If the `onChange` prop is specified, the component enters controlled mode, and the user
+ * is responsible for handling the `onChange` event, as well as providing the `selectedTab`
+ * 
+ * If not, the component, will automatically handle dispatching events to the Redux store
+ */
 export default function SideContentProvider({
   tabs,
   defaultTab,
   children,
-  ...props
+  onChange,
+  selectedTab: propsSelectedTab,
+  workspaceLocation,
 }: SideContentProviderProps) {
   const { alerts, height, dynamicTabs, setSelectedTab, selectedTab } = useSideContent(
-    props.workspaceLocation,
+    workspaceLocation,
     defaultTab
   );
+
+  useEffect(() => {
+    if (propsSelectedTab === undefined && onChange) {
+      console.warn('onChange was provided, but selectedTab was not. Changes to selectedTab will not be reflected')
+    }
+  }, [onChange, propsSelectedTab])
 
   const allTabs = tabs
     ? [...tabs.beforeDynamicTabs, ...dynamicTabs, ...tabs.afterDynamicTabs]
     : dynamicTabs;
 
   const changeTabsCallback: ChangeTabsCallback = (newId, oldId, event) => {
-    if (props.onChange) {
+    if (onChange) {
       // Controlled mode
-      props.onChange(newId, oldId, event);
+      onChange(newId, oldId, event);
     } else {
       // Uncontrolled mode
       setSelectedTab(newId);
@@ -58,7 +81,7 @@ export default function SideContentProvider({
     tabs: allTabs,
     alerts,
     changeTabsCallback,
-    selectedTab: props.selectedTab ? props.selectedTab : selectedTab,
+    selectedTab: propsSelectedTab ?? selectedTab,
     height
   });
 }
