+++ src/features/stories/StoriesTypes.ts
@@ -26,10 +26,6 @@ export const GET_STORIES_USER = 'GET_STORIES_USER';
 export const SET_CURRENT_STORIES_USER = 'SET_CURRENT_STORIES_USER';
 export const SET_CURRENT_STORIES_GROUP = 'SET_CURRENT_STORIES_GROUP';
 
-export const STORIES_VISIT_SIDE_CONTENT = 'STORIES_VISIT_SIDE_CONTENT';
-export const BEGIN_STORIES_ALERT_SIDE_CONTENT = 'BEGIN_STORIES_ALERT_SIDE_CONTENT';
-export const END_STORIES_ALERT_SIDE_CONTENT = 'END_STORIES_ALERT_SIDE_CONTENT';
-
 export type StoryMetadata = {
   authorId: number;
   authorName: string;
